fix: AbstractUIData.findComponent() patch
issue: TOBAGO-2154
